+++ src/components/pricing/PurchaseTeamsTier.js
@@ -1,20 +1,10 @@
 import React, { useState } from 'react';
 import { Button, NumberOfEngineers, HelpText } from 'components';
+import { OPTIONS_FOR_NUMBER_OF_ENGINEERS } from 'components/forms/NumberOfEngineers';
 
 import TierName from './TierName';
 import TierDescription from './TierDescription';
 import TierBulletsSection from './TierBulletsSection';
-
-export const OPTIONS_FOR_NUMBER_OF_ENGINEERS = [
-  Object.freeze({ id: 10, name: '10 developers', dollarCentCostPerDevPerMonth: 1800 }),
-  Object.freeze({ id: 20, name: '20 developers', dollarCentCostPerDevPerMonth: 1800 }),
-  Object.freeze({ id: 30, name: '30 developers', dollarCentCostPerDevPerMonth: 1800 }),
-  Object.freeze({ id: 40, name: '40 developers', dollarCentCostPerDevPerMonth: 1800 }),
-  Object.freeze({ id: 50, name: '50 developers', dollarCentCostPerDevPerMonth: 1600 }),
-  Object.freeze({ id: 75, name: '75 developers', dollarCentCostPerDevPerMonth: 1600 }),
-  Object.freeze({ id: 100, name: '100 developers', dollarCentCostPerDevPerMonth: 1500 }),
-  Object.freeze({ id: 150, name: '150 developers', dollarCentCostPerDevPerMonth: 1500 }),
-];
 
 const PurchaseTeamsTier = () => {
   const [numberOfEngineers, setNumberOfEngineers] = useState(OPTIONS_FOR_NUMBER_OF_ENGINEERS[0].id);
@@ -56,17 +46,19 @@ const PurchaseTeamsTier = () => {
 
         <div>
           <div className="mb-1">
+
             <Button
-              text="Start a free trial"
-              link={true}
+              text="Buy now"
               color="primary"
-              to="/free-trial/"
+              data-cb-type="checkout"
+              data-cb-item-0="Roadie-Teams-Plan-USD-Monthly"
+              data-cb-item-0-quantity="1"
               fullWidth
             />
           </div>
 
           <div className="text-center">
-            <HelpText message="No credit card required" />
+            <HelpText message="Powered by Stripe and Chargebee" />
           </div>
         </div>
       </div>
