+++ src/components/pricing/TeamsPricingTier.js
@@ -1,16 +1,11 @@
 import React, { useState } from 'react';
 import { Button, NumberOfEngineers, HelpText } from 'components';
 
+import { OPTIONS_FOR_NUMBER_OF_ENGINEERS } from 'components/forms/NumberOfEngineers';
+
 import TierName from './TierName';
 import TierDescription from './TierDescription';
 import TierBulletsSection from './TierBulletsSection';
-
-export const OPTIONS_FOR_NUMBER_OF_ENGINEERS = [
-  Object.freeze({ id: 50, name: '50 developers', dollarCentCostPerDevPerMonth: 1800 }),
-  Object.freeze({ id: 70, name: '70 developers', dollarCentCostPerDevPerMonth: 1700 }),
-  Object.freeze({ id: 100, name: '100 developers', dollarCentCostPerDevPerMonth: 1600 }),
-  Object.freeze({ id: 150, name: '150 developers', dollarCentCostPerDevPerMonth: 1500 }),
-];
 
 const TeamsPricingTier = () => {
   const [numberOfEngineers, setNumberOfEngineers] = useState(OPTIONS_FOR_NUMBER_OF_ENGINEERS[0].id);
